Break out test files into test folders.
